- Adding in support for better caching of modules
Now the node modules will be cached during the build process most of the time. If the NPM or Bower dependencies change, a new build will need to be done.  This should take care of 99% of the cases in a lab environment, only requring the NPM Install once.

If we wanted to get more clever, we could also push the result of the build to the Hub and then change to use that image